requirements.txt
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.5
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
firebase-admin==5.0.3
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
googlemaps==4.5.3
grpcio==1.41.0
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.20.1
idna==3.2
jsonpickle==2.0.0
msgpack==1.0.2
packaging==21.0
proto-plus==1.19.5
protobuf==3.18.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyspellchecker==0.6.2
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.6
whitenoise==5.3.0
